# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Go Post Article",
        "title": "Go Post Article",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/article": {
            "get": {
                "description": "Retrieve a paginated list of posts with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "List posts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (publish, draft, thrash)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page limit (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns {posts: []entity.Post, total: int}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a post with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Create a new post",
                "parameters": [
                    {
                        "description": "Create Post Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreatePost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Post"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation error",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    }
                }
            }
        },
        "/article/{id}": {
            "get": {
                "description": "Retrieve details of a single post using its unique ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get a post by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Post"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a post from the system by ID",
                "tags": [
                    "posts"
                ],
                "summary": "Delete a post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update existing post details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Update a post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Post Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdatePost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Post"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Post": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "Programming"
                },
                "content": {
                    "type": "string",
                    "example": "The content of the article goes here..."
                },
                "created_date": {
                    "type": "string",
                    "example": "2026-05-14T11:38:11Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_nabilfikrisp_sv-be_internal_entity.PostStatus"
                        }
                    ],
                    "example": "Publish"
                },
                "title": {
                    "type": "string",
                    "example": "How to Learn Go"
                },
                "updated_date": {
                    "type": "string",
                    "example": "2026-05-14T11:38:11Z"
                }
            }
        },
        "github_com_nabilfikrisp_sv-be_internal_entity.PostStatus": {
            "type": "string",
            "enum": [
                "publish",
                "draft",
                "thrash"
            ],
            "x-enum-varnames": [
                "StatusPublish",
                "StatusDraft",
                "StatusThrash"
            ]
        },
        "v1.CreatePost": {
            "type": "object",
            "required": [
                "category",
                "content",
                "status",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "programming"
                },
                "content": {
                    "type": "string",
                    "minLength": 200,
                    "example": "Lorem ipsum dolor sit amet..."
                },
                "status": {
                    "enum": [
                        "publish",
                        "draft",
                        "thrash"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_nabilfikrisp_sv-be_internal_entity.PostStatus"
                        }
                    ],
                    "example": "publish"
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 20,
                    "example": "how to learn go programming"
                }
            }
        },
        "v1.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "v1.UpdatePost": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "updated category"
                },
                "content": {
                    "type": "string",
                    "minLength": 200,
                    "example": "updated content..."
                },
                "status": {
                    "enum": [
                        "publish",
                        "draft",
                        "thrash"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_nabilfikrisp_sv-be_internal_entity.PostStatus"
                        }
                    ],
                    "example": "draft"
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 20,
                    "example": "updated title"
                }
            }
        }
    }
}